// src/api/types.gen.ts
export type CustomDataFieldBase = {
    /**
     * Name
     */
    name: string;
    /**
     * Can User Answer
     */
    can_user_answer: boolean;
};
// ...
export type GenerateTicketBase = {
    /**
     * Name
     */
    name: string;
    /**
     * Max Use
     */
    max_use: number;
    /**
     * Expiration
     */
    expiration: string;
};